backend/requirements.txt
fastapi
uvicorn
torch
diffusers
rembg
pillow
